Rename base_test_case to test_base_case.
Setuptools sdist packages test/test*.py files with your distribution by default,
allowing for execution of setup.py test.

It wasn't grabbing test/base_test_case.py but it will grab test/test_case_base.py.
