+++ test/test_flaky_decorator.py
@@ -3,7 +3,7 @@
 from __future__ import unicode_literals
 from flaky.flaky_decorator import flaky
 from flaky.names import FlakyNames
-from test.base_test_case import TestCase
+from test.test_case_base import TestCase
 
 
 class TestFlakyDecorator(TestCase):
